fixed broken viewport test, populated documentation in confusing sections of the codebase, added initial position setting of mesh entities

+++ product/ViewPortController.h
@@ -37,5 +37,5 @@ public:
 	~ViewPortController(){}
 
 	std::vector<float> GetCameraAngle();
-	std::vector<int> GetViewPortDimensions();
+	std::vector<float> GetViewPortDimensions();
 };
